Make presentation adapter generic so it can be reused
