+++ EssentialApp/EssentialApp/LoadResourcePresentationAdapter.swift
@@ -1,0 +1,34 @@
+import EssentialDeveloper
+import EssentialDeveloperiOS
+import Combine
+
+final class LoadResourcePresentationAdapter<Resource, View: ResourceView> {
+    private let loader: () -> AnyPublisher<Resource, Error>
+    private var cancellable: Cancellable?
+    var presenter: LoadResourcePresenter<Resource, View>?
+    
+    init(loader: @escaping () -> AnyPublisher<Resource, Error>) {
+        self.loader = loader
+    }
+    
+    func loadResource() {
+        presenter?.didStartLoading()
+        
+        cancellable = loader().sink { [weak self] completion in
+            switch completion {
+            case .finished:
+                break
+            case let .failure(error):
+                self?.presenter?.didFinishLoading(with: error)
+            }
+        } receiveValue: { [weak self] resource in
+            self?.presenter?.didFinishLoading(with: resource)
+        }
+    }
+}
+
+extension LoadResourcePresentationAdapter: FeedViewControllerDelegate {
+    func didRequestFeedRefresh() {
+        loadResource()
+    }
+}
